+++ CHANGELOG.md
@@ -20,7 +20,8 @@ All notable changes to Jodd project are documented here.
 + **db** - fixed `LocalDate` and `LocalDateTime` mappings.
 + **http** - allowing `contentTypeJson` to be used in any order.
 + **http** - fixed an issue with re-reading the request.
-+ **http** - fixed an issue with query param parsing. 
++ **http** - fixed an issue with query param parsing.
++ **core** - fixed up Java version parsing.
 
 ## [v5.0.6](https://github.com/oblac/jodd/compare/v5.0.5...master)